backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python-headless
numpy
torch
torchvision
segment-anything
albumentations
scikit-learn
mediapipe
pillow
